+++ Flow/src/utopia/flow/generic/model/immutable/Constant.scala
@@ -1,10 +1,12 @@
 package utopia.flow.generic.model.immutable
 
-import utopia.flow.generic.model.mutable.Variable
-import utopia.flow.generic.model.template.Property
+import utopia.flow.collection.immutable.{Pair, PairView}
+import utopia.flow.generic.model.mutable.{DataType, Variable}
+import utopia.flow.generic.model.template.{Property, ValueConvertible}
 import utopia.flow.operator.equality.EqualsExtensions._
-import utopia.flow.operator.equality.ApproxEquals
+import utopia.flow.operator.equality.{ApproxEquals, EqualsBy}
 import utopia.flow.util.logging.Logger
+import utopia.flow.view.immutable.caching.Lazy
 
 import scala.language.implicitConversions
 
@@ -22,6 +24,34 @@ object Constant
 	// OTHER    -------------------------
 	
 	/**
+	 * @param name Name of this constant
+	 * @param value Value of this constant
+	 * @return A new constant with the specified name & value
+	 */
+	def apply(name: String, value: Value): Constant = _Constant(name, value)
+	
+	/**
+	 * @param name Name of this constant
+	 * @param value A lazy container that will yield the value for this constant
+	 * @return A new constant with the specified name & value
+	 */
+	def lazily(name: String, value: Lazy[Value]): Constant = value.current match {
+		case Some(value) => apply(name, value)
+		case None => new LazyConstant(name, value)
+	}
+	/**
+	 * @param name Name of this constant
+	 * @param getValue A function for acquiring the value of this constant
+	 *                 (in some form convertible to a [[ValueConvertible]] instance)
+	 * @param toValue Implicit value conversion for the value type.
+	 *                Usually from [[utopia.flow.generic.casting.ValueConversions]].
+	 * @tparam V Type of the value yielded by 'getValue'
+	 * @return A constant where the value is initialized lazily
+	 */
+	def lazily[V](name: String)(getValue: => V)(implicit toValue: V => ValueConvertible): Constant =
+		new LazyConstant(name, Lazy { toValue(getValue).toValue })
+	
+	/**
 	 * @param property A property
 	 * @return A constant with the same name & (current) value as that property
 	 */
@@ -29,6 +59,25 @@ object Constant
 		case c: Constant => c
 		case p => apply(p.name, p.value)
 	}
+	
+	
+	// NESTED   -------------------------
+	
+	private class LazyConstant(override val name: String, lazyValue: Lazy[Value]) extends Constant
+	{
+		// IMPLEMENTED  -----------------
+		
+		override def value: Value = lazyValue.value
+		
+		override protected def equalsProperties: IterableOnce[Any] = PairView(name, value)
+		
+		override def withName(name: String): Constant = lazyValue.current match {
+			case Some(value) => Constant(name, value)
+			case None => new LazyConstant(name, lazyValue)
+		}
+	}
+	
+	private case class _Constant(name: String, value: Value) extends Constant
 }
 
 /**
@@ -36,13 +85,8 @@ object Constant
   * @author Mikko Hilpinen
   * @since 29.11.2016
   */
-case class Constant(name: String, value: Value) extends Property with ApproxEquals[Property]
+trait Constant extends Property with EqualsBy with ApproxEquals[Property]
 {
-	// COMP. PROPERTIES    ---------
-	
-	override val dataType = value.dataType
-	
-	
 	// COMPUTED    -----------------
 	
 	/**
@@ -54,6 +98,10 @@ case class Constant(name: String, value: Value) extends Property with ApproxEqua
 	
 	// IMPLEMENTED  ----------------
 	
+	override def dataType: DataType = value.dataType
+	
+	override protected def equalsProperties: IterableOnce[Any] = Pair(name, value)
+	
 	override def ~==(other: Property): Boolean = (name ~== other.name) && (value ~== other.value)
 	
 	
@@ -63,12 +111,12 @@ case class Constant(name: String, value: Value) extends Property with ApproxEqua
 	  * Creates a new constant that has the provided value but the same name
 	  * @param value the value the new constant will have
 	  */
-	def withValue(value: Value) = copy(value = value)
+	def withValue(value: Value) = Constant(name, value)
 	/**
 	  * @param name New name for constant
 	  * @return A copy of this constant with provided name
 	  */
-	def withName(name: String) = copy(name = name)
+	def withName(name: String) = Constant(name, value)
 	
 	/**
 	  * @param f A mapping function for this constant's value